refactor(Introduction/Sidebar): file structure changes
